Initial work on incorporating CodeMirror
Status: The html loads into a dialog, and there is syntax highlighting.
However, there are errors reported from the CodeMirror js files because
the initial preload and instantiation of CodeMirror does not yet happen.

Followed http://www.avoid.org/codemirror-for-tinymce4/ for general
approach.

Followed suit on how the kottiimage_plugin is appended to the
kotti_tinymce resource_mapping in __init__.py, and is configured in
richtext.pt.

Added the code button per the instructions in the link, and removed all
but the minimum configuration of CodeMirror.

Modified static/plugins/codemirror/plugin.js to load the code at
/codemirrorsource, which is simply configured to load all the js files
with .need() calls in an __init__ config.

Perhaps the solution is to initialize CodeMirror in a manner similar to
how TinyMCE is pre-loaded and initialized in richtext.pt?